feat: optimize data acquisition logic
